test(analysis): add syntax tests for shielded types in ABI encoding
Add tests verifying the compiler rejects shielded types in:
- abi.encode() with suint256, saddress, sbool, structs, and arrays
- abi.encodePacked() with shielded types
- abi.encodeWithSelector() and abi.encodeWithSignature()
- abi.decode() with shielded type targets

Update existing tests to expect the new error and cast shielded NONCE
to uint96 in AESEncryptAndDecryptWithRng semantic test.

These tests expose a vulnerability where shielded values could be
serialized to plaintext bytes via ABI encoding, leaking confidential
data and allowing recovery via abi.decode.

+++ test/libsolidity/semanticTests/functionCall/seismic_precompiles/AESEncryptAndDecryptWithRng.sol
@@ -20,7 +20,7 @@ contract AES {
     function encryptAndStore() public {
         address AESEncryptAddr = address(0x66);
 
-        bytes memory input = abi.encodePacked(AES_KEY, NONCE, PLAINTEXT);
+        bytes memory input = abi.encodePacked(AES_KEY, uint96(NONCE), PLAINTEXT);
 
         (bool success, bytes memory output) = AESEncryptAddr.staticcall(input);
         require(success, "Precompile encryption call failed");
@@ -34,7 +34,7 @@ contract AES {
 
         address AESDecryptAddr = address(0x67);
 
-        bytes memory input = abi.encodePacked(AES_KEY, NONCE, encryptedData);
+        bytes memory input = abi.encodePacked(AES_KEY, uint96(NONCE), encryptedData);
 
         (bool success, bytes memory output) = AESDecryptAddr.staticcall(input);
         require(success, "Precompile decryption call failed");
